% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blankSub.R
\name{blankSub}
\alias{blankSub}
\title{substract blanks (i.e. negative controls/ background) from peak table}
\usage{
blankSub(peakTable = NULL, blankNames = NULL, sampNames = NULL,
  method = "mean", thresh = 2)
}
\arguments{
\item{peakTable}{either a data.frame, full file path as a character string to a  .csv file of a peak table in the form observation (samples) in columns and
variables (Mass spectral signals) in rows. If argument is not supplied a GUI file selection window will open and a .csv file can be selected.}

\item{blankNames}{character vector of blank (i.e. negative control) names to identify
appropriate observation (blank) columns.}

\item{sampNames}{character vector of sample names to identify appropriate observation (sample) columns.}

\item{method}{character either 'median' or 'mean' fold change calculation.}

\item{thresh}{numeric sample:blank fold change cut-off. Any LC-MS
features below this threshold will be removed. (default = 2).}
}
\value{
a data frame similar to peakTable with any rows (LC-MS features) removed
which are below the minimum blank fold change threshold.
}
\description{
calculate either the mean or median fold change between samples
and blanks for every LC-MS variable and retain only LC-MS features above
the blank fold change cut-off.
}
