% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvCalc.R
\name{cvCalc}
\alias{cvCalc}
\title{calculate the analytical coefficient of variation from a set of observations}
\usage{
cvCalc(peakTable = NULL, obsNames = NULL, thresh = 30)
}
\arguments{
\item{peakTable}{either a data.frame, full file path as a character string to a  .csv file of a peak table in the form observation (samples) in columns and
variables (Mass spectral signals) in rows. If argument is not supplied a GUI (tcltk) file selection window will open and a .csv file can be selected.}

\item{obsNames}{character vector (i.e. sample/ QC/ Blank) names to identify appropriate observation (sample) columns
with to calculate the coefficient of variance (CV\%).}

\item{thresh}{numeric the minimum CV\% to retain an LC-MS variable (default = 30 i.e. 30\%).}
}
\value{
a data frame similar to peakTable with any rows (LC-MS features) removed
which are above the maximum CV\% threshold. In addition a new column is added
'coeffVar' which contains the CV\% calculated for each variable.
}
\description{
calculate the coefficient of variation (CV\%) for all LC-MS features
of peak table using a set of  observations (e.g. pooled quality control samples).
The CV\% is calculated according to the following equation:

\eqn{CV\% = (\delta / \mu) * 100}

where \eqn{\delta} is the standard deviation and \eqn{\mu} is the mean
}

