% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoScapeMap.R
\name{cytoScapeMap}
\alias{cytoScapeMap}
\title{Generate cytoscape map files from peakTable}
\usage{
cytoScapeMap(peakTable = NULL, obsNames = NULL, outDir = NULL,
  EICnos = NULL, fileNameId = "", corrThresh = 0.6,
  corrMethod = "spearman", delta = 0.05, MTC = "BH")
}
\arguments{
\item{peakTable}{either a data.frame, full file path as a character string to a  .csv file of a peak table in the form observation (samples) in columns and
variables (Mass spectral signals) in rows. If argument is not supplied a GUI file selection window will open and a .csv file can be selected.}

\item{obsNames}{character vector of observation (i.e. sample/ QC/ Blank) names to identify appropriate observation (sample) columns.}

\item{outDir}{character full path of cytoscape file output directory. The function
will create an additional subdirectory within this output called "cytoScapeFiles" into
which files will be saved.}

\item{EICnos}{character name of column in peakTable of unique feature numbers. 
This unique number (e.g. extracted ion chromatogram number) will be used to 
merge the peak table with the node attributes file.}

\item{fileNameId}{character additional identificatory character string to 
add to the cytoscape output file names. default = "". e.g. "blankSubtracted"
or "filtered".}

\item{corrThresh}{correlation coefficient threshold to group features within
a retention time cluster.}

\item{corrMethod}{character correlation method see \code{\link{cor}} for details. default "spearman".}

\item{delta}{numeric maximum p-value (following multiple testing correction) above 
which the null hypothesis (no correlation) is rejected.}

\item{MTC}{character Multiple Testing Correction default is "BH", see \code{\link{p.adjust.methods}} for
details of options. ("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr", "none").
Any p-values after multiple testing correction above the value of delta will have their
corresponding correlation coefficents replaced with zero.}
}
\value{
writes 3 text files: network (.sif), node and edge attribute files as 
plain text (.txt) files in the output directory (see outDir).
}
\description{
Generate cytoscape map files from peakTable
}
