% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hotEllipse.R
\name{hotEllipse}
\alias{hotEllipse}
\title{Hotellings T2 statistic and ellipse calculation function taken from
pcaMethods:::simpleEllipse}
\usage{
hotEllipse(x, y, alfa = 0.95, len = 10000)
}
\description{
Hotellings T2 statistic and ellipse calculation function taken from
pcaMethods:::simpleEllipse
}

