% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loessFitKfoldCV.R
\name{loessFitKfoldCV}
\alias{loessFitKfoldCV}
\title{loess with 7 - fold CV function}
\usage{
loessFitKfoldCV(x, y, feature, folds = 7)
}
\arguments{
\item{x}{predictor values}

\item{y}{response values}

\item{folds}{numeric (default=7, i.e. 7-fold cross validation).}

\item{original}{LC-MS feature.}
}
\description{
loess with 7 - fold CV function
}

