% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loessSmooth.R
\name{loessSmooth}
\alias{loessSmooth}
\title{pooled quality control based signal drift/ attenuation smoothing}
\source{
\url{http://www.nature.com/nprot/journal/v6/n7/abs/nprot.2011.335.html} 
Procedures for large-scale metabolic profiling of serum and plasma using gas 
chromatography and liquid chromatography coupled to mass spectrometry
}
\usage{
loessSmooth(peakTable = NULL, sampNames = NULL, qcNames = NULL,
  nCores = NULL, outputDir = NULL, smoothSpan = NULL, folds = 7)
}
\arguments{
\item{peakTable}{either a data.frame, full file path as a character string to a  .csv file of a peak table in the form observation (samples) in columns and
variables (Mass spectral signals) in rows. If argument is not supplied a GUI file selection window will open and a .csv file can be selected.}

\item{sampNames}{character vector of sample names to identify appropriate observation (sample) columns.}

\item{qcNames}{character vector of quality control (QC) names to identify appropriate observation (sample) columns.}

\item{nCores}{number of cores to use for parallel processing with the foreach
and snow packages.}

\item{outputDir}{optional directory path to save output images before and after QC
smoothing. A subdirectory will be created in which to save the png images.}

\item{smoothSpan}{fixed smoothing span. If supplied a this fixed smoothing
span parameter will override the cross validated feature-by-feature smoothing
span optimization.}

\item{folds}{numeric (default=7, i.e. 7-fold cross validation).}
}
\value{
a data frame identical to peakTable with signal drift/ attenuation adjusted.
}
\description{
attempts to correct for systematic drifts in LC-MS datasets using
intervally injected pooled quality control samples representative of all samples
of a experiment.
}
\details{
systematic drifts in mass spectral signal can be corrected using
intervally injected pooled quality control samples for each mass spectral signal
variable in the peak table. The optimum span parameter for each LC-MS feature 
is identified using 7-fold cross-validation 
(see: \code{\link{crossval}} from the bootstrap package). A new column is
added to the peakTable "smoothSpanLoessFit" containing the result of the 
optimal value of span for the loess function. 
 For each LC-MS variable \code{\link{loess}} models
of the quality control data are calculated using direct surface fitting (see: \code{\link{loess.control}}) and
the loess fit using 7-fold cross validation is carried out by a slight
modification of the \code{\link{loess.wrapper}} function from the \code{\link{bisoreg}}
package. 
The optimum loess model is then used to predict (see: \code{\link{predict.loess}}) 
intensity values for both the QC and sample injections according
to the degree of smoothing (span) in the final optimum loess model. 
The product of the original LC-MS variable intensity and the deviation of the 
loess predicted values from their median is then calculated to adjust the 
signal drift. 
This function attempts to
adjust systematic drift/ attenuation within each LC-MS feature. 
The degree of smoothing determined by the smoothSpan argument however results
may vary slightly using the K-fold cross validation method. In order to maintain
absolute reproducibility it is recommend to set the seed prior to using this 
function (see: \code{\link{set.seed}}). It is recommended that parallel processing
is considered to reduce the computational time cost using the argument nCores.
}
\seealso{
\code{\link{loess}}, \code{\link{loess.wrapper}}, \code{\link{bisoreg}},
 \code{\link{crossval}}, \code{\link{predict.loess}}, \code{\link{set.seed}}.
}
