% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loessWrapperMod.R
\name{loessWrapperMod}
\alias{loessWrapperMod}
\title{adapted from \link{\code{bisoreg}} package}
\usage{
loessWrapperMod(x, y, data, span.vals = seq(0.25, 1, by = 0.05), folds = 5)
}
\description{
adapted from \link{\code{bisoreg}} package
}

