% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/logTrans.R
\name{logTrans}
\alias{logTrans}
\title{log transform peak table}
\usage{
logTrans(peakTable = NULL, obsNames = NULL, base = exp(1))
}
\arguments{
\item{peakTable}{either a data.frame, full file path as a character string to a  .csv file of a peak table in the form observation (samples) in columns and
variables (Mass spectral signals) in rows. If argument is not supplied a GUI file selection window will open and a .csv file can be selected.}

\item{obsNames}{character vector of observation (i.e. sample/ QC/ Blank) names to identify appropriate observation (sample) columns.}

\item{base}{the base with respect to which logarithms are computed \code{\link{log}}. Defaults to e=exp(1)}
}
\value{
a data frame identical to peakTable with observations log transformed.
}
\description{
log transform zero filled peak table to correct for multiplicative noise.
}

