% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lowessSmooth.R
\name{lowessSmooth}
\alias{lowessSmooth}
\title{pooled quality control based signal drift/ attenuation smoothing}
\source{
\url{http://www.nature.com/nprot/journal/v6/n7/abs/nprot.2011.335.html}
Procedures for large-scale metabolic profiling of serum and plasma using gas
chromatography and liquid chromatography coupled to mass spectrometry
}
\usage{
lowessSmooth(peakTable = NULL, sampNames = NULL, qcNames = NULL,
  smoothSpan = 0.2, ...)
}
\arguments{
\item{peakTable}{either a data.frame, full file path as a character string to a  .csv file of a peak table in the form observation (samples) in columns and
variables (Mass spectral signals) in rows. If argument is not supplied a GUI file selection window will open and a .csv file can be selected.}

\item{sampNames}{character vector of sample names to identify appropriate observation (sample) columns.}

\item{qcNames}{character vector of quality control (QC) names to identify appropriate observation (sample) columns.}

\item{smoothSpan}{smoother span, proportion of total data points used for smoothing. See \code{\link{lowess}}. default = 0.2}

\item{...}{additional arguments to \code{\link{lowess}}.}
}
\value{
a data frame identical to peakTable with signal drift/ attenuation adjusted.
}
\description{
attempts to correct for systematic drifts in LC-MS datasets using
intervally injected pooled quality control samples representative of all samples
of a experiment.
}
\details{
systematic drifts in mass spectral signal can be corrected using
intervally injected pooled quality control samples for each mass spectral signal
variable in the peak table. Sample intensities are first replaced
with NAs and cubic spline interpolation (\code{\link{na.spline}} from the zoo package)
is used to replace missing values between QC samples. The \code{\link{lowess}} function
is then used to model systematic drift/ attenuation within each mass spectral
signal variable. The degree of smoothing determined by the smoothSpan argument.
}
\seealso{
\code{\link{lowess}}, \code{\link{na.spline}}.
}

