% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattanPlot.R
\name{manhattanPlot}
\alias{manhattanPlot}
\title{manhattan plot function}
\usage{
manhattanPlot(x = NULL, y = NULL, data = NULL, cols = NULL,
  plotOutputFile = NULL, minPval = 0.05, labelFontSize = 15, ...)
}
\arguments{
\item{x}{character string of a single column name representing the x axis (e.g. order or mass-to charge ratio).}

\item{y}{character string of a single column name (if data argument supplied) containing the statistical test p-values the y axis.}

\item{data}{data.frame containing the x and y columns}

\item{cols}{numeric vector of integers, hexadecimal colours or names of colours for data points.}

\item{plotOutputFile}{character string full path to write output image file (see\link{\code{ggsave}}).}

\item{minPval}{numeric minimum p-value to plot significance line and change point shapes.}

\item{labelFontSize}{numeric label and title font size (supplied to \link{\code{theme_bw}})}

\item{...}{additional arguments to \link{\code{ggsave}}.}
}
\description{
manhattan plot function
}
