% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaClustId.R
\name{pcaClustId}
\alias{pcaClustId}
\title{Identify clusters in a PCA plot from a list of co-variates using PAM clustering.}
\usage{
pcaClustId(pcaResult = NULL, peakTable = NULL, coVarTable = NULL,
  obsNames = NULL, ...)
}
\arguments{
\item{pcaResult}{a \code{\link{pcaRes}} class object}

\item{peakTable}{optional if pcaResult argument not supplied. Either a data.frame, full file path as a character string to a  .csv file of a peak table in the form observation (samples) in columns and
variables (Mass spectral signals) in rows. If argument is not supplied a GUI file selection window will open and a .csv file can be selected.}

\item{coVarTable}{either a data.frame, full file path as a character string to a  .csv file of a co-variates table in the form observation (sample) names in the first column and
co-variates from the 2nd column onward. If argument is not supplied a GUI file selection window will open and a .csv file can be selected.}

\item{obsNames}{character vector of observation (i.e. sample/ QC/ Blank) names to identify appropriate observation (sample) columns.}

\item{...}{additional arguments to \code{\link{pamk}}.}
}
\value{
a list containing three elements:

1. a list pamkClust containing three elements returned from \code{\link{pamk}}

2. a list lmCoVarClust containing the linear models obtained from \code{\link{lm}}.

3. a named character vector rSquaredLm containing the coefficients of determination
from the linear models. Named with each covariate or two-factor interaction
considered.
}
\description{
attempts to the identify clusters in a pca from a data frame
of covariates, using partitioning around the medoid clustering.
}
\details{
potential clusters in \code{\link{pcaRes}} scores are identified using partitioning
around the medoid clustering (\code{\link{pamk}}) from the fpc package with an
estimation of the number of clusters. Given a data.frame of co-variates/ sample information, 
the most likely explanatory co-variate AND/OR potential two-factor interactions will be established using linear modelling 
\code{\link{lm}}.

Co-variates containing missing values or only one unique value will not be considered.
The linear model consists of response ~ terms where response is the clusters
established by \code{\link{pamk}} and terms are the factor levels of the
co-variate table. The best explanatory co-variate for the PCA clustering 
is defined as the linear model with the highest coefficient of determination (R2).
}
\seealso{
\code{\link{lowess}}, \code{\link{na.spline}}, \code{\link{pcaOutId}}.
}
