% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaResOut.R
\name{pcaResOut}
\alias{pcaResOut}
\title{expand hotellings ellipse from a pcaRes object \code{\link{pcaRes}}
and expand hotellings ellipse to identify outliers}
\usage{
pcaResOut(pcaResult, outTol)
}
\arguments{
\item{pcaResult}{a \code{\link{pcaRes}} object.}

\item{outTol}{proportional expansion value for Hotelling's ellipse (PC1 and PC2), 
any outlying samples beyond this will be removed.}
}
\value{
list containing pca results file, and named logical vector of possible
outliers.
}
\description{
expand hotellings ellipse from a pcaRes object \code{\link{pcaRes}}
and expand hotellings ellipse to identify outliers
}

