% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPcsEx.R
\name{plotPcsEx}
\alias{plotPcsEx}
\title{plot PCA with expanded Hotelling's ellipse adapted from \code{\link{plotPcs}} from 
the \code{\link{pcaMethods}} package.}
\usage{
plotPcsEx(object, exEllipse, type = c("scores", "loadings"), ...)
}
\arguments{
\item{object}{\code{\link{pcaRes}} class object}

\item{exEllipse}{matrix of expanded Hotelling's Ellipse see \code{\link{pcaOutId}}.}

\item{type}{character either "scores" or "loadings" for scores or loadings plot respectively}

\item{Further}{arguments to \code{\link{pairs}} on which this function is based.}
}
\description{
plot PCA with expanded Hotelling's ellipse adapted from \code{\link{plotPcs}} from 
the \code{\link{pcaMethods}} package.
}
\seealso{
\code{\link{pcaOutId}}, \code{\link{pcaRes}}, \code{\link{pca}}.
}

