% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ramClustMod.R
\name{ramClustMod}
\alias{ramClustMod}
\title{\code{\link{ramclustR}} - modified for MetMSLine}
\source{
\url{http://pubs.acs.org/doi/abs/10.1021/ac501530d} RAMClust: A Novel 
Feature Clustering Method Enables Spectral-Matching-Based Annotation for 
Metabolomics Data.
}
\usage{
ramClustMod(peakTable = NULL, obsNames = NULL, st = 3, sr = 0.5,
  maxt = 20, mult = 5, deepSplit = FALSE, blocksize = 2000,
  hmax = 0.3, minModuleSize = 2, linkage = "average")
}
\arguments{
\item{peakTable}{optional if pcaResult argument not supplied. Either a data.frame, 
full file path as a character string to a  .csv file of a peak table in the 
form observation (samples) in columns and variables (Mass spectral signals) 
in rows. If argument is not supplied a GUI file selection window will open 
and a .csv file can be selected.}

\item{obsNames}{character vector of observation (i.e. sample/ QC/ Blank) names 
to identify appropriate observation (sample) columns.}

\item{st}{Sigma t value controlling the rate at which feature similarity 
decays with respect to differences in retention times between feature pairs. 
When an xcms diff report is used as input, this parameter is by default set to 
half the peak width of all the peaks in the xcmsSet. Can be manually overridden 
to tune performance when necessary. Larger values are less selective, smaller 
values are more selective.}

\item{sr}{Sigma r value controlling the rate at which feature similarity 
decays with respect to correlational simiarity between a pair of features. 
numeric e.g. 0.5, Sigma r: the rate of similarity decay in response to 
correlational r . Can be manually overridden to tune performance. Larger 
values are less selective, smaller values are more selective.}

\item{maxt}{numeric e.g 20, the maximum time allowed for similarity between 
features to be greater than zero. For example, if two features are separated 
by more than 20 seconds, the similarity score is automatically set to zero, 
ensuring that they end up in different clusters. If you are importing data 
from another vendor, ensure that either you change the feature times to 
seconds (default in ramclustR) or you change this value to a value 
representing minutes, rather than seconds (e.g. 0.33). Used primarily to 
speed up compuation time by not calculating values above this threshold.}

\item{deepSplit}{boolean e.g. =FALSE, access to the deepSplit function in 
\code{\link{dynamicTreeCut}}. False generates fewer spectra, 
TRUE splits spectra more readily and thereby generates more spectra. 
For more detailed description, see documentation for the 
\code{\link{dynamicTreeCut}} package.}

\item{blocksize}{integer number of features processed in one block =2000, 
ramclustR makes attempts to minimize memory demands to maintain processing 
speed by avoiding memory constraints. It does so by processing the full 
similarity matrix in blocks of blocksize dimensions. This has two benefits: 
one is that a given PC is less likely to run out of memory during processing 
and the second is that we need not process blocks that contain not feature 
pairs with a delta rt of greater than maxt, thereby avoiding needless 
calculations and processing time.}

\item{minModuleSize}{the number of features required for a cluster to be 
reported. That is, if a feature is clustered in a group with less than 
minModuleSize features, it will not be exported. If you wish to analyze 
singletons, set minModuleSize=1.}

\item{linkage}{The linkage method used to perform heirarchical 
fastcluster-based heirarchical clustering. default="average".}
}
\value{
a list containing 2 elements: 

 1. the original peak table with the results of retention time 
 correlation clustering included in additional column "RtCorrClust" as a 
 data frame.
 
 2. A data frame of the weighted means and details of the most intense 
 feature for each group. In addition for each group details of the best sample 
 to consider for subsequent MSn fragmentation analysis based on the most 
 top 5 most intense samples is provided. This "best" sample for MS/MS is also
 selected whilst concurrently minimising the number of samples to potentially 
 reinject for targetted MSn fragmentation analysis.
}
\description{
\code{\link{ramclustR}} - modified for MetMSLine
}

