% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtCorrClust.R
\name{rtCorrClust}
\alias{rtCorrClust}
\title{rtCorrClust}
\usage{
rtCorrClust(peakTable = NULL, obsNames = NULL, rtThresh = NULL,
  corrThresh = 0.6, minFeat = 2, hclustMethod = "median",
  distMeas = "euclidean", corMethod = "spearman")
}
\arguments{
\item{peakTable}{either a data.frame, full file path as a character string to a  .csv file of a peak table in the form observation (samples) in columns and
variables (Mass spectral signals) in rows. If argument is not supplied a GUI file selection window will open and a .csv file can be selected.}

\item{obsNames}{character vector of observation (i.e. sample/ QC/ Blank) names to identify appropriate observation (sample) columns.}

\item{rtThresh}{retention time threshold for retention time clustering.}

\item{corrThresh}{correlation coefficient threshold (non-parametric Spearman's 
Rho) to group features within a retention time cluster.}

\item{minFeat}{minimum number of features with a Retention time/ correlation
cluster to consider it a group (default = 2, i.e. a cluster must contain at
least 2 features to be considered a group).}

\item{hclustMethod}{hierarchical clustering method to \code{\link{hclust.vector}} 
method of fastcluster package (default = "median").}

\item{distMeas}{distance measure for retention time clustering (default = "euclidean"). 
see \code{\link{hclust.vector}}.}

\item{corMethod}{correlation method to \code{\link{cor}} (default = "spearman).
Within retention clusters dissimilarity is computed as 1-correlation coefficient.}
}
\value{
a list containing 2 elements: 

 1. the original peak table with the results of retention time 
 correlation clustering included in additional columns "rtgroup", "CorrClusts", 
"RtCorrClust" as a data frame and details of the most intense 
 feature for each group. In addition for each group details of the best sample 
 to consider for subsequent MSn fragmentation analysis based on the most 
 top 5 most intense samples is provided. This "best" sample for MS/MS is also
 selected whilst concurrently minimising the number of samples to potentially 
 reinject for targetted MSn fragmentation analysis.
 
 2. A data frame of the weighted means and details of the most intense 
 feature for each group. In addition for each group details of the best sample 
 to consider for subsequent MSn fragmentation analysis based on the most 
 top 5 most intense samples is provided. This "best" sample for MS/MS is also
 selected whilst concurrently minimising the number of samples to potentially 
 reinject for targetted MSn fragmentation analysis. 
  @seealso \code{\link{hclust.vector}}, \code{\link{cutree}}.
}
\description{
performs retention time clustering thin intra-retention time cluster
correlation coefficient (spearman's Rho) clustering.
}
\details{
The \code{\link{cutree}} function is used to identify retention time clusters based on a cut height
equal to retention time (rtThresh) supplied.
Within retention time cluster correlation clustering is based on 1 - correlation
coefficient dissimilarity. The cut height of intra retention time cluster
correlation clusters is based on a minimum correlation coefficient threshold
(default = 0.6). Following retention time/ correlation cluster identification
any groups containing less than a mimimum feature number are removed 
(minFeat default = 2), a default value for the minFeat of 2 means that only
isolated single features with no relation to another feature are removed.
}
