% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signNorm.R
\name{signNorm}
\alias{signNorm}
\title{Normalize mass spectral data peak table}
\source{
\url{http://www.ncbi.nlm.nih.gov/pubmed/21526840} 
Optimized preprocessing of ultra-performance liquid chromatography/mass spectrometry 
urinary metabolic profiles for improved information recovery.
}
\usage{
signNorm(peakTable = NULL, obsNames = NULL, method = "medFC")
}
\arguments{
\item{peakTable}{either a data.frame, full file path as a character string to a  .csv file of a peak table in the form observation (samples) in columns and
variables (Mass spectral signals) in rows. If argument is not supplied a GUI file selection window will open and a .csv file can be selected.}

\item{obsNames}{character vector of observation (i.e. sample/ QC/ Blank) names to identify appropriate observation (sample) columns.}

\item{method}{either "medFC" for median fold change or "totIon" for total ion signal normalization.
also a custom vector of factors equal in length to the obsNames argument
 with which to normalize the data can also be supplied. default = "medFC".}
}
\value{
a data frame identical to peakTable with samples normalized according to specificied method.
}
\description{
normalizes mass spectral signal either by the median fold change 
(probabilistic quotient) or total ion signal methods
}

