% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableCheckRead.R
\name{tableCheckRead}
\alias{tableCheckRead}
\title{peak/ co-variate table error handling and/ or read from character string}
\usage{
tableCheckRead(tmpTable = NULL, type = "peak-picker", ...)
}
\arguments{
\item{tmpTable}{either a data.frame, full file path as a character string to a text file (.csv/.txt/.tsv) of a table 
If argument is not supplied a GUI file selection window will open and a text file (.csv/.txt/.tsv) can be selected.}

\item{type}{description of table type as character string for gui windows and error messages.}

\item{...}{additional arguments to \code{\link{fread}}.}
}
\description{
peak/ co-variate table error handling and/ or read from character string
}

