% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/targMS2csv.R
\name{targMS2csv}
\alias{targMS2csv}
\title{Generate target list (.csv) file from subset peak table}
\usage{
targMS2csv(peakTable = NULL, filename = NULL, manufacturer = "Agilent",
  deltaMz = 20, deltaRt = 1, isoWidth = c("Narrow (~1.3 m/z)",
  "Medium (~4 m/z)"))
}
\arguments{
\item{peakTable}{either a data.frame, full file path as a character string to a  .csv file of a peak table in the form observation (samples) in columns and
variables (Mass spectral signals) in rows. If argument is not supplied a GUI file selection window will open and a .csv file can be selected.}

\item{filename}{character full path file name to save the target list file to.}

\item{deltaMz}{numeric delta mass accuracy (ppm) for precursor selection.}

\item{deltaRt}{numeric delta retention time in minutes}

\item{isoWidth}{character isolation width for quadrupole. either 'Narrow (~1.3 m/z)'
or 'Medium (~4 m/z)'.}
}
\description{
Generate target list (.csv) file from subset peak table
}

