% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoPlot.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Volcano plot slightly modified from \code{\link{VolcanoPlot}} from the metabolomics
package.}
\usage{
volcanoPlot(folds, pvals, cexcutoff = 0.7, cexlab = 0.5, plimit = 0.05,
  fclimit = 2, xlab = "log2 Fold Change", ylab = "-log10 t-Test P-value",
  main = "Volcano Plot", ...)
}
\arguments{
\item{folds}{A vector of fold changes with metabolite names.}

\item{pvals}{A vector of corresponding p-values with metabolite names.}

\item{cexcutoff}{Font size of the cut-off labels.}

\item{cexlab}{Font size of the variable labels.}

\item{plimit}{A numeric indicating the p value cutoff. The default is set to 0.05.}

\item{fclimit}{A numeric indicating the lower fold cutoff. The default is set to 2.}

\item{xlab}{x-axis label.}

\item{ylab}{y-axis label}

\item{main}{Plot title.}

\item{...}{Other graphical parameters. See par.}
}
\description{
Volcano plot slightly modified from \code{\link{VolcanoPlot}} from the metabolomics
package.
}
