% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zeroFill.R
\name{zeroFill}
\alias{zeroFill}
\title{zero fill peak table}
\usage{
zeroFill(peakTable = NULL, obsNames = NULL, value = NULL)
}
\arguments{
\item{peakTable}{either a data.frame, full file path as a character string to a  .csv file of a peak table in the form observation (samples) in columns and
variables (Mass spectral signals) in rows. If argument is not supplied a GUI file selection window will open and a .csv file can be selected.}

\item{obsNames}{character vector of observation (i.e. sample/ QC/ Blank) names to identify appropriate observation (sample) columns.}

\item{value}{numeric value to fill zero/ missing values (NA). If left blank
then half the mimimum non-zero observed peak intensity is used.}
}
\value{
a data frame identical to peakTable with zero/ NAs filled with half
the mimimum non-zero observed value.
}
\description{
identify and replace zero/ missing values (NA) in peak table with half
the mimimum non-zero observed value.
}

